requirements/development.txt
-r base.txt
django-debug-toolbar==4.2.*
django-extensions==3.2.*
pytest==7.4.*
pytest-django==4.5.*
factory-boy==3.3.* 